using light intensity

+++ src/rt_lights/components.rs
@@ -4,7 +4,7 @@ use bevy::prelude::*;
 
 #[derive(Component, Default, Clone, Copy)]
 pub struct PointLight {
-    pub intensity: u8,
+    pub intensity: f32,
     pub colour: Color
 }
